Refactored and tested LoanAgreementApplicationService

+++ src/PipefittersAccounting.Infrastructure/Application/Commands/Financing/LoanAgreementAggregate/LoanAgreementDeleteCommand.cs
@@ -1,8 +1,10 @@
 using PipefittersAccounting.Core.Financing.LoanAgreementAggregate;
 using PipefittersAccounting.Core.Financing.LoanAgreementAggregate.ValueObjects;
-using PipefittersAccounting.Infrastructure.Interfaces;
 using PipefittersAccounting.Core.Interfaces.Financing;
+using PipefittersAccounting.Infrastructure.Interfaces;
+using PipefittersAccounting.Infrastructure.Interfaces.Financing;
 using PipefittersAccounting.SharedKernel.CommonValueObjects;
+using PipefittersAccounting.SharedKernel;
 using PipefittersAccounting.SharedKernel.Utilities;
 using PipefittersAccounting.SharedModel.WriteModels.Financing;
 
@@ -10,10 +12,11 @@ namespace PipefittersAccounting.Infrastructure.Application.Commands.Financing.Lo
 {
     public class LoanAgreementDeleteCommand
     {
-        public static async Task<OperationResult<bool>> Execute
+        public static async Task<OperationResult<bool>> Process
         (
-            DeleteLoanAgreementInfo model,
+            LoanAgreementWriteModel model,
             ILoanAgreementAggregateRepository repo,
+            ILoanAgreementValidationService validationService,
             IUnitOfWork unitOfWork
         )
         {
@@ -26,17 +29,26 @@ namespace PipefittersAccounting.Infrastructure.Application.Commands.Financing.Lo
                     return OperationResult<bool>.CreateFailure(errMsg);
                 }
 
-                OperationResult<LoanAgreement> result = await repo.GetByIdAsync(model.LoanId);
+                ValidationResult validationResult = await validationService.IsValidDeleteLoanAgreementInfo(model);
 
-                if (!result.Success)
+                if (validationResult.IsValid)
                 {
-                    return OperationResult<bool>.CreateFailure(result.NonSuccessMessage);
+                    OperationResult<LoanAgreement> result = await repo.GetByIdAsync(model.LoanId);
+
+                    if (!result.Success)
+                    {
+                        return OperationResult<bool>.CreateFailure(result.NonSuccessMessage);
+                    }
+
+                    repo.Delete(result.Result);
+                    await unitOfWork.Commit();
+
+                    return OperationResult<bool>.CreateSuccessResult(true);
                 }
-
-                repo.Delete(result.Result);
-                await unitOfWork.Commit();
-
-                return OperationResult<bool>.CreateSuccessResult(true);
+                else
+                {
+                    return OperationResult<bool>.CreateFailure(validationResult.Messages[0]);
+                }
             }
             catch (Exception ex)
             {
